Make fixes for API changes, fly.io. email freq
The casa bonita api has changed. It now requires the headers I've included and removal of the fc_token param. The response format has also changed so I've updated the parsing of that accordingly. Also they once in a while return a captcha which breaks the res.json() call so if the app notices that then you'll get an email notifying you to solve the captcha which will unblock the bot until the next time a captcha gets returned.

I've updated the fly.toml so the machine doesn't auto shut off during the sleep cycle. And also reduce the size of the machine to the fee tier size.

Sendgrid free tier only allows 100 emails a day so I reduced the frequency of emails being sent out to comply with that.